change forcast to get data and forecastItem to store data
